member_portal/requirements-new.txt
# ═════════════════════════════════════════════════════════
# TNT Alliance Portal - Python Dependencies
# ═════════════════════════════════════════════════════════

# ── Web Framework ──
aiohttp>=3.11.18
Jinja2>=3.1.4

# ── Database ──
# SQLite3 مدمج في Python

# ── Configuration ──
python-dotenv>=1.1.0

# ── Security ──
bcrypt>=4.1.0  # اختياري إذا أردت استخدام bcrypt بدلاً من PBKDF2

# ── Email ──
# smtplib مدمج في Python

# ── OAuth2 ──
aiohttp-oauth>=0.1.0  # اختياري

# ── Utilities ──
certifi>=2025.4.26
requests>=2.32.3  # للطلبات العادية

# ── Development ──
pytest>=7.0.0
pytest-aiohttp>=1.3.0
black>=22.0.0
flake8>=4.0.0
mypy>=1.0.0

# ── Logging ──
# logging مدمج في Python

# ── Optional But Recommended ──
asyncio-contextmanager>=1.0.0
